refactor: separate logic to container in create, edit page
