+++ src/pages/CreatePost/CreatePost.tsx
@@ -1,5 +1,4 @@
-import { useRecoilValue } from 'recoil';
-import axios from 'axios';
+import React from 'react';
 import { Button, getAccessToken } from '../../shared';
 import { CreateFormView, MapView } from './views';
 import {
@@ -11,95 +10,24 @@ import {
   StSkyline,
   StGrayBg,
 } from './style';
-import { useMutation } from '@tanstack/react-query';
-import { useNavigate } from 'react-router-dom';
-
-import { createPostState } from './state';
 import { Theme } from '../../styles';
-import { EModalType, useModal } from '../../providers';
+import { useCreatePost } from './container';
 
 export const CreatePost = () => {
-  const { openModal, closeModal } = useModal();
   const token = getAccessToken();
-  const post = useRecoilValue(createPostState);
-  const navigate = useNavigate();
 
   if (!token) {
     window.location.replace('javascript:history.back()');
     return null;
   }
 
-  const isFormValid = () => {
-    return (
-      post.title &&
-      post.content &&
-      post.address &&
-      post.latitude &&
-      post.longitude &&
-      post.images &&
-      post.images.length > 0
-    );
-  };
+  const handlers = useCreatePost();
 
-  const handleCreatePostClick = () => {
-    if (isFormValid()) {
-      openModal(EModalType.POP_UP, {
-        title: '작성한 게시물을 업로드 할까요?',
-        cancelButton: true,
-        functionButton: {
-          theme: 'blue',
-          label: '확인',
-          onClick: () => {
-            mutation.mutate(), closeModal();
-          },
-        },
-      });
-    } else {
-      openModal(EModalType.POP_UP, {
-        title: '양식을 모두 입력해주세요',
-        cancelButton: false,
-        functionButton: {
-          theme: 'emptyBlue',
-          label: '확인',
-          onClick: () => closeModal(),
-        },
-      });
-    }
-  };
+  if (!handlers) {
+    return null;
+  }
 
-  console.log(post);
-
-  const sendPostRequest = async () => {
-    const formData = new FormData();
-    const postJSON = JSON.stringify({
-      title: post.title,
-      content: post.content,
-      latitude: post.latitude,
-      longitude: post.longitude,
-      address: post.address,
-    });
-
-    const blob = new Blob([postJSON], { type: 'application/json' });
-    formData.append('post', blob);
-    post.images.forEach((image) => {
-      formData.append(`images`, image);
-    });
-
-    return axios.post(`${process.env.REACT_APP_API_URI}/api/posts`, formData, {
-      headers: {
-        Authorization: `Bearer ${token}`,
-      },
-    });
-  };
-
-  const mutation = useMutation(sendPostRequest, {
-    onSuccess: (data) => {
-      navigate(`/detail/${data.data.data.postId}`);
-    },
-    onError: (error) => {
-      console.error('게시 요청 실패:', error);
-    },
-  });
+  const { handleCreatePostClick, handleCancelClick } = handlers;
 
   return (
     <StCreatePostContainer>
@@ -112,20 +40,7 @@ export const CreatePost = () => {
             $buttonTheme="empty"
             size="mediumLong"
             fontSize={Theme.fontSizes.h2}
-            onClick={() =>
-              openModal(EModalType.POP_UP, {
-                title: '작업을 중단하고 나가시겠습니까?',
-                cancelButton: true,
-                functionButton: {
-                  theme: 'pink',
-                  label: '나가기',
-                  onClick: () => {
-                    navigate(-1);
-                    closeModal();
-                  },
-                },
-              })
-            }
+            onClick={handleCancelClick}
           />
         </StBtnBox1>
         <StBtnBox2>
@@ -138,7 +53,6 @@ export const CreatePost = () => {
           />
         </StBtnBox2>
       </StBtnContainer>
-
       <StBg>
         <StSkyline></StSkyline>
         <StGrayBg></StGrayBg>
