+++ src/pages/CreatePost/container/index.ts
@@ -1,1 +1,3 @@
-export {};
+export * from './CreateFormContainer';
+export * from './CreatePostContainer';
+export * from './MapContainer';
