+++ src/pages/CreatePost/views/CreateFormView/CreateFormView.tsx
@@ -11,45 +11,18 @@ import {
   StPhotoBoxInput,
   StCloseButton,
 } from './style';
-import React, { useState } from 'react';
 import { HelpIcon } from '../../../../assets';
-import { useRecoilState } from 'recoil';
-import { createPostState } from '../../state';
+import { useCreateForm } from '../../container';
 
 export const CreateFormView = () => {
-  const [post, setPost] = useRecoilState(createPostState);
-  const [previewImages, setPreviewImages] = useState<string[]>([]);
-
-  const handleImageUpload = (e: React.ChangeEvent<HTMLInputElement>) => {
-    const files = e.target.files;
-    if (files) {
-      const fileArray = Array.from(files);
-      const imagesArray = fileArray.map((file) => URL.createObjectURL(file));
-
-      const newPreviewImages = [...previewImages, ...imagesArray].slice(0, 3);
-      const newFiles = [...(post?.images || []), ...fileArray].slice(0, 3);
-
-      setPreviewImages(newPreviewImages);
-      setPost({ ...post, images: newFiles });
-    }
-  };
-
-  const removeImage = (index: number) => {
-    const updatedImages = [...previewImages];
-    const updatedImgs = [...post.images];
-    updatedImages.splice(index, 1);
-    updatedImgs.splice(index, 1);
-    setPreviewImages(updatedImages);
-    setPost({ ...post, images: updatedImgs });
-  };
-
-  const handleTitleChange = (e: React.ChangeEvent<HTMLInputElement>) => {
-    setPost({ ...post, title: e.target.value });
-  };
-
-  const handleContentChange = (e: React.ChangeEvent<HTMLTextAreaElement>) => {
-    setPost({ ...post, content: e.target.value });
-  };
+  const {
+    post,
+    previewImages,
+    handleImageUpload,
+    removeImage,
+    handleTitleChange,
+    handleContentChange,
+  } = useCreateForm();
 
   return (
     <>
@@ -66,7 +39,7 @@ export const CreateFormView = () => {
         <StContentContainer>
           <StTextContainer>
             <textarea
-              placeholder="민원 내용을 작성해주세요 *"
+              placeholder="내용을 작성해주세요 *"
               value={post.content}
               onChange={handleContentChange}
               maxLength={500}
