+++ src/pages/CreatePost/views/MapView/MapView.tsx
@@ -1,7 +1,8 @@
-import { useEffect, useRef, useState } from 'react';
+import { useRef } from 'react';
 import { useRecoilState } from 'recoil';
-import { MyDirectIcon, PingIcon } from '../../../../assets';
+import { MyDirectIcon } from '../../../../assets';
 import { MapIcon } from '../../../../shared/MapIcon';
+import { useMap } from '../../container';
 import { createPostState } from '../../state';
 import {
   StContainer,
@@ -21,106 +22,26 @@ declare global {
 
 export const MapView = () => {
   const mapContainerRef = useRef<HTMLDivElement | null>(null);
-  const [map, setMap] = useState<any>(null);
   const [post, setPost] = useRecoilState(createPostState);
-  const [zoomLevel, setZoomLevel] = useState<number>(3);
-  const [mapCenter, setMapCenter] = useState({
+
+  const initialMapCenter = {
     lat: 37.565869791860365,
     lng: 126.98258019375905,
-  });
-
-  useEffect(() => {
-    if (!map && window.kakao && window.kakao.maps && mapContainerRef.current) {
-      const { Map, services, event, Marker, InfoWindow, MarkerImage } =
-        window.kakao.maps;
-      const options = {
-        center: new window.kakao.maps.LatLng(mapCenter.lat, mapCenter.lng),
-        level: zoomLevel,
-      };
-
-      const mapInstance = new Map(mapContainerRef.current, options);
-      setMap(mapInstance);
-
-      const geocoder = new services.Geocoder();
-      const pingIconSrc = PingIcon(false);
-      const imageSize = new window.kakao.maps.Size(40, 40);
-      const imageOption = { offset: new window.kakao.maps.Point(20, 20) };
-      const markerImage = new MarkerImage(pingIconSrc, imageSize, imageOption);
-      const marker = new Marker({ image: markerImage });
-
-      event.addListener(mapInstance, 'click', (mouseEvent: any) => {
-        searchDetailAddrFromCoords(
-          geocoder,
-          mouseEvent.latLng,
-          (result: any, status: any) => {
-            if (status === services.Status.OK) {
-              const address = result[0].address.address_name;
-              const latitude = mouseEvent.latLng.getLat();
-              const longitude = mouseEvent.latLng.getLng();
-              setPost((prevPost) => ({
-                ...prevPost,
-                address,
-                latitude,
-                longitude,
-              }));
-
-              marker.setPosition(mouseEvent.latLng);
-              marker.setMap(mapInstance);
-            }
-          },
-        );
-      });
-    }
-  }, []);
-
-  useEffect(() => {
-    if (map) {
-      map.setLevel(zoomLevel);
-    }
-  }, [zoomLevel, map]);
-
-  useEffect(() => {
-    if (map) {
-      const newCenter = new window.kakao.maps.LatLng(
-        mapCenter.lat,
-        mapCenter.lng,
-      );
-      map.setCenter(newCenter);
-    }
-  }, [mapCenter, map]);
-
-  const searchDetailAddrFromCoords = (
-    geocoder: any,
-    coords: any,
-    callback: any,
-  ) => {
-    geocoder.coord2Address(coords.getLng(), coords.getLat(), callback);
   };
+  const initialZoomLevel = 3;
 
-  const zoomChangeHandler = (newZoomLevel: number) => {
-    setZoomLevel(newZoomLevel);
-
-    if (map) {
-      const currentCenter = map.getCenter();
-      const newMapCenter = {
-        lat: currentCenter.getLat(),
-        lng: currentCenter.getLng(),
-      };
-      mapCenterChangeHandler(newMapCenter);
-    }
-  };
-
-  const mapCenterChangeHandler = (newMapCenter: {
-    lat: number;
-    lng: number;
-  }) => {
-    setMapCenter(newMapCenter);
-  };
+  const {
+    map,
+    zoomLevel,
+    mapCenter,
+    zoomChangeHandler,
+    mapCenterChangeHandler,
+  } = useMap(mapContainerRef, initialMapCenter, initialZoomLevel);
 
   return (
     <>
       <StContainer>
-        <StMapText>민원 위치 설정하기 *</StMapText>
+        <StMapText>사건 위치 설정하기 *</StMapText>
         <StMapContainer ref={mapContainerRef}>
           <StMainMapIcon>
             <MapIcon
