+++ src/pages/EditPost/EditPost.tsx
@@ -1,7 +1,3 @@
-import axios from 'axios';
-import { useMutation, useQuery } from '@tanstack/react-query';
-import { useState, useEffect } from 'react';
-import { useNavigate, useParams } from 'react-router-dom';
 import { EditFormView } from './views';
 import {
   StEditPostContainer,
@@ -12,71 +8,17 @@ import {
   StSkyline,
   StGrayBg,
 } from './style';
-import { Button, getAccessToken } from '../../shared';
-import { EditPostType } from './type';
 import { Theme } from '../../styles';
 import { EditMapView } from './views/EditMapView/EditMapView';
 import { EModalType, useModal } from '../../providers';
+import { Button } from '../../shared';
+import { useNavigate } from 'react-router-dom';
+import { useEditPost } from './container';
 
 export const EditPost = () => {
   const { openModal, closeModal } = useModal();
-  const token = getAccessToken();
-  const [post, setPost] = useState<EditPostType | null>(null);
-
+  const { post, setPost, isLoading, isError, mutation } = useEditPost();
   const navigate = useNavigate();
-
-  const { id: stringId } = useParams();
-  const id = Number(stringId);
-
-  const getPostDetail = async (id: number) => {
-    const headers = token
-      ? {
-          Authorization: `Bearer ${token}`,
-        }
-      : undefined;
-
-    const response = await axios.get(
-      `${process.env.REACT_APP_API_URI}/api/posts/${id}`,
-      {
-        headers: headers,
-      },
-    );
-    return response.data.data;
-  };
-
-  const { data, isError, isLoading } = useQuery(['postDetail', id], () =>
-    getPostDetail(id),
-  );
-
-  console.log(data);
-
-  if (data && !data.owner) {
-    window.location.replace('javascript:history.back()');
-    return null;
-  }
-
-  useEffect(() => {
-    if (data) {
-      setPost(data);
-    }
-  }, [data, setPost]);
-
-  const mutation = useMutation(
-    () =>
-      axios.put(`${process.env.REACT_APP_API_URI}/api/posts/${id}`, post, {
-        headers: {
-          Authorization: `Bearer ${token}`,
-        },
-      }),
-    {
-      onSuccess: () => {
-        navigate(`/detail/${id}`);
-      },
-      onError: (error) => {
-        console.error('수정 요청 실패:', error);
-      },
-    },
-  );
 
   if (isLoading) {
     return <div>Loading...</div>;
