+++ src/pages/EditPost/container/index.ts
@@ -1,0 +1,2 @@
+export * from './EditPostContainer';
+export * from './EditFormContainer';
