+++ src/pages/EditPost/views/index.ts
@@ -1,1 +1,2 @@
 export * from './EditFormView/EditFormView';
+export * from './EditMapView/EditMapView';
